+++ TabletopTweaks/Bugfixes/Classes/Lich.cs
@@ -1,11 +1,14 @@
 ﻿using HarmonyLib;
+using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.ElementsSystem;
 using Kingmaker.RuleSystem;
 using Kingmaker.RuleSystem.Rules.Damage;
 using Kingmaker.UnitLogic.Mechanics;
 using Kingmaker.UnitLogic.Mechanics.Components;
+using System.Linq;
 using TabletopTweaks.Config;
 using TabletopTweaks.Extensions;
 
@@ -21,6 +24,7 @@ namespace TabletopTweaks.Bugfixes.Classes {
                 Main.LogHeader("Patching Lich Resources");
 
                 PatchDeathRush();
+                PatchSpellbookMerging();
 
                 void PatchDeathRush() {
                     if (ModSettings.Fixes.Lich.IsDisabled("DeathRush")) { return; }
@@ -44,6 +48,23 @@ namespace TabletopTweaks.Bugfixes.Classes {
                         };
                     });
                     Main.LogPatch("Patched", DeathRushFeature);
+                }
+
+                void PatchSpellbookMerging() {
+
+                    if (ModSettings.Fixes.Lich.IsDisabled("SpellbookMerging")) { return; }
+
+                    var LichIncorporateSpellbookFeature = Resources.GetBlueprint<BlueprintFeatureSelectMythicSpellbook>("3f16e9caf7c683c40884c7c455ed26af");
+                    var ExploiterWizardSpellbook = Resources.GetBlueprint<BlueprintSpellbook>("d09794fb6f93e4a40929a965b434070d");
+                    var NatureMageSpellbook = Resources.GetBlueprint<BlueprintSpellbook>("3ed7e38dc8134af28e1a2b105f74fb7b");
+
+                    LichIncorporateSpellbookFeature.m_AllowedSpellbooks = LichIncorporateSpellbookFeature.m_AllowedSpellbooks
+                        .AddItem(ExploiterWizardSpellbook.ToReference<BlueprintSpellbookReference>())
+                        .AddItem(NatureMageSpellbook.ToReference<BlueprintSpellbookReference>())
+                        .Distinct()
+                        .ToArray();
+
+                    Main.LogPatch("Patched", LichIncorporateSpellbookFeature);
                 }
             }
         }